Altered the initialisation of yfronts
Restructured and commented out yfronts.

Made a little bit progress formatting a selection without using the
built in execcommand stuff as the markup it outputs is crap.

Added in a tastey CSS animation for sliding in the toolbar when a
contenteditable field is focussed.

I also added in some logic to hide the toolbar when it is tabbed out of.